Added change password form to docs
