+++ src/js/main.js
@@ -3,6 +3,7 @@ import './modules/polyfill-remove';
 import requestForm from './modules/submission.js'
 import swagger from './modules/swagger.js'
 import bulkSubmission from './modules/bulk-submission.js'
+import changePassword from './modules/change-password.js'
 import requestFormJSON from './modules/request-form-json.js'
 import Icons from 'nyco-patterns/dist/elements/icons/Icons.common'
 import Toggle from 'nyco-patterns/dist/utilities/toggle/Toggle.common'
@@ -26,6 +27,9 @@ if ((window.location.pathname.indexOf('request-builder') >= 0)) {
 }
 if ((window.location.pathname.indexOf('bulk-submission') >= 0)) {
   bulkSubmission();
+}
+if ((window.location.pathname.indexOf('change-password') >= 0)) {
+  chnagePassword();
 }
 
 /* Get the content markdown from CDN and append */
