+++ src/views/change-password.twig
@@ -1,0 +1,46 @@
+{% extends "layouts/default.twig" %}
+
+{% set title = 'Change Password Form' %}
+{% set page_title = title %}
+
+{% block content %}
+<p>The usage of this page requires an account. You can <a class="{{ classes.link }}" href="http://eepurl.com/gfLTuH" target="_blank">request an account here</a>. See <a href="getting-started">Getting Started</a> for more details.</p>
+
+<form action="https://screeningapi.cityofnewyork.us/" method="post" novalidate="true" data-js="change-password" class="border border-solid border-primary-blue-30t bg-secondary-white-70t p-4">
+
+  <div id="c_username" class="mb-2">
+    <label class="font-bold" for="username">Username <span class="text-primary-red">*</span></label>
+
+    <div>
+      <input type="text" id="username" placeholder="Enter Username" name="username" maxlength="255" required="true">
+    </div>
+  </div>
+
+  <div id="c_password" class="mb-2">
+    <label class="font-bold" for="password">Password <span class="text-primary-red">*</span></label>
+
+    <div>
+      <input type="password" id="password" placeholder="Password" name="password" maxlength="255" required="true">
+    </div>
+  </div>
+
+  <div class="mb-2">
+      <label class="font-bold" for="newpassword">New Password</label>
+      <div>
+          <input class="mt-1" type="password" id="newpassword" placeholder="New Password" name="newPassword" maxlength="255">
+      </div>
+  </div>
+
+  <div class="flex items-center">
+    <button type="submit" name="submit" class="btn color-primary-button">Submit</button>
+
+    <div style="display: none;" id="loader">
+      <div class="flex items-center px-2 text-primary-blue">
+        {% include 'partials/spinner.twig' %} &nbsp;&nbsp; Loading...
+      </div>
+    </div>
+  </div>
+
+  <div id="errors" role="alert" class="p-3 mt-2 bg-primary-red-30t border border-primary-red" style="display: none"></div>
+</form>
+{% endblock %}
