+++ src/views/partials/nav.twig
@@ -11,6 +11,7 @@
 
   <span class="{{ classes.header }}">Tools</span>
   <a class="{{ classes.link }}" href="request-builder">Request Builder</a>
+  <a class="{{ classes.link }}" href="change-password">Change Password</a>
   <a class="{{ classes.link }}" href="bulk-submission">Bulk Submission Form</a>
 
   <span class="{{ classes.header }}">Reference</span>
